--- pkg/corerp/frontend/controller/applications/deleteapplication.go
@@ -1,75 +1,0 @@
-/*
-Copyright 2023 The Radius Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package applications
-
-import (
-	"context"
-	"errors"
-	"net/http"
-
-	v1 "github.com/project-radius/radius/pkg/armrpc/api/v1"
-	ctrl "github.com/project-radius/radius/pkg/armrpc/frontend/controller"
-	"github.com/project-radius/radius/pkg/armrpc/rest"
-	"github.com/project-radius/radius/pkg/corerp/datamodel"
-	"github.com/project-radius/radius/pkg/corerp/datamodel/converter"
-	"github.com/project-radius/radius/pkg/ucp/store"
-)
-
-var _ ctrl.Controller = (*DeleteApplication)(nil)
-
-// DeleteApplication is the controller implementation to delete application resource.
-type DeleteApplication struct {
-	ctrl.Operation[*datamodel.Application, datamodel.Application]
-}
-
-// NewDeleteApplication creates a new DeleteApplication.
-func NewDeleteApplication(opts ctrl.Options) (ctrl.Controller, error) {
-	return &DeleteApplication{
-		ctrl.NewOperation(opts,
-			ctrl.ResourceOptions[datamodel.Application]{
-				RequestConverter:  converter.ApplicationDataModelFromVersioned,
-				ResponseConverter: converter.ApplicationDataModelToVersioned,
-			},
-		),
-	}, nil
-}
-
-func (a *DeleteApplication) Run(ctx context.Context, w http.ResponseWriter, req *http.Request) (rest.Response, error) {
-	serviceCtx := v1.ARMRequestContextFromContext(ctx)
-
-	old, etag, err := a.GetResource(ctx, serviceCtx.ResourceID)
-	if err != nil {
-		return nil, err
-	}
-
-	if old == nil {
-		return rest.NewNoContentResponse(), nil
-	}
-
-	if r, err := a.PrepareResource(ctx, req, nil, old, etag); r != nil || err != nil {
-		return r, err
-	}
-
-	if err := a.StorageClient().Delete(ctx, serviceCtx.ResourceID.String()); err != nil {
-		if errors.Is(&store.ErrNotFound{}, err) {
-			return rest.NewNoContentResponse(), nil
-		}
-		return nil, err
-	}
-
-	return rest.NewOKResponse(nil), nil
-}
