+++ pkg/corerp/frontend/controller/containers/validator.go
@@ -25,7 +25,7 @@ import (
 )
 
 // ValidateAndMutateRequest validates and mutates the incoming request.
-func ValidateAndMutateRequest(ctx context.Context, newResource *datamodel.ContainerResource, oldResource *datamodel.ContainerResource, options *controller.Options) (rest.Response, error) {
+func ValidateAndMutateRequest(ctx context.Context, newResource, oldResource *datamodel.ContainerResource, options *controller.Options) (rest.Response, error) {
 	if newResource.Properties.Identity != nil {
 		return rest.NewBadRequestResponse("User-defined identity in Applications.Core/containers is not supported."), nil
 	}
