--- pkg/corerp/frontend/controller/gateways/testdata/requestheaders20220315privatepreview.json
@@ -1,63 +1,0 @@
-{
-    "Accept": "application/json",
-    "Accept-Encoding": "gzip, deflate",
-    "Accept-Language": "en-US",
-    "Content-Length": "305",
-    "Content-Type": "application/json; charset=utf-8",
-    "Referer": "https://radius.dev/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/radius-test-rg/providers/applications.core/applications/app0?api-version=2022-03-15-privatepreview",
-    "Traceparent": "00-000011048df2134ca37c9a689c3a0000-0000000000000000-01",
-    "User-Agent": "ARMClient/1.6.0.0",
-    "Via": "1.1 Azure",
-    "X-Azure-Requestchain": "hops=1",
-    "X-Fd-Clienthttpversion": "1.1",
-    "X-Fd-Clientip": "0000:0000:0000:1:0000:0000:0000:0000",
-    "X-Fd-Edgeenvironment": "fake",
-    "X-Fd-Eventid": "00005A12DDEC4F8B80B65BB768190000",
-    "X-Fd-Impressionguid": "00005A12DDEC4F8B80B65BB768190000",
-    "X-Fd-Originalurl": "https://radius.dev:443/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/radius-test-rg/providers/Applications.Core/applications/app0?api-version=2022-03-15-privatepreview",
-    "X-Fd-Partner": "AzureResourceManager_Test",
-    "X-Fd-Ref": "Ref A: xxxx Ref B: xxxx Ref C: 2022-03-22T18:54:50Z",
-    "X-Fd-Revip": "country=United States,iso=us,state=Washington,city=Redmond,zip=00000,tz=-8,asn=0,lat=0,long=-1,countrycf=8,citycf=8",
-    "X-Fd-Routekey": "000075000",
-    "X-Fd-Socketip": "0000:0000:0000:1:0000:0000:0000:0000",
-    "X-Forwarded-For": "192.168.0.10",
-    "X-Forwarded-Host": "radius.dev",
-    "X-Forwarded-Port": "443",
-    "X-Forwarded-Proto": "https",
-    "X-Forwarded-Scheme": "https",
-    "X-Ms-Activity-Vector": "IN.0P",
-    "X-Ms-Arm-Network-Source": "PublicNetwork",
-    "X-Ms-Arm-Request-Tracking-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Arm-Resource-System-Data": "{\"lastModifiedBy\":\"fake@hotmail.com\",\"lastModifiedByType\":\"User\",\"lastModifiedAt\":\"2022-03-22T18:57:52.6857175Z\"}",
-    "X-Ms-Arm-Service-Request-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Client-Acr": "1",
-    "X-Ms-Client-Alt-Sec-Id": "1:live.com:0006000017E40000",
-    "X-Ms-Client-App-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Client-App-Id-Acr": "0",
-    "X-Ms-Client-Audience": "https://management.core.windows.net/",
-    "X-Ms-Client-Authentication-Methods": "pwd",
-    "X-Ms-Client-Authorization-Source": "RoleBased",
-    "X-Ms-Client-Family-Name-Encoded": "fake",
-    "X-Ms-Client-Given-Name-Encoded": "fake",
-    "X-Ms-Client-Identity-Provider": "live.com",
-    "X-Ms-Client-Ip-Address": "192.168.0.10",
-    "X-Ms-Client-Issuer": "https://sts.windows-ppe.net/00000000-0000-0000-0000-000000000000/",
-    "X-Ms-Client-Location": "centralus",
-    "X-Ms-Client-Object-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Client-Principal-Group-Membership-Source": "Token",
-    "X-Ms-Client-Principal-Id": "000000000000000",
-    "X-Ms-Client-Principal-Name": "live.com#fake@hotmail.com",
-    "X-Ms-Client-Puid": "000000000000000",
-    "X-Ms-Client-Request-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Client-Scope": "user_impersonation",
-    "X-Ms-Client-Tenant-Id": "00000000-0000-0000-0000-000000000001",
-    "X-Ms-Client-Wids": "00000000-0000-0000-0000-000000000000, 00000000-0000-0000-0000-000000000001",
-    "X-Ms-Correlation-Request-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Home-Tenant-Id": "00000000-0000-0000-0000-000000000002",
-    "X-Ms-Request-Id": "00000000-0000-0000-0000-000000000000",
-    "X-Ms-Routing-Request-Id": "CENTRALUS:20220322T185452Z:00000000-0000-0000-0000-000000000000",
-    "X-Original-Forwarded-For": "0000:0000:0000:1:449b:f928:e40a:a351",
-    "X-Real-Ip": "192.168.0.10",
-    "X-Request-Id": "1000f6040000000000004bc7d1666424",
-    "X-Scheme": "https"
-}
